Node style and build tree from scratch
Changed the style of the node to a rectangle with an operator on the
left and some content on the bottom-right corner. Also, implemented the
functionality to build the tree from scratch. When the add button is
clicked, a new node will appear on the right which you can drag and
drop to the tree.